hnu-min/requirements.txt
fastapi
uvicorn
chromadb
crawl4ai
cssselect
openai>=1.0.0
apscheduler
python-dotenv
streamlit
requests
